[FLINK-38779] Move TableIdRouter to flink-cdc-common package
